refactor: remove barrel utils
