+++ packages/runner/src/utils/tasks.ts
@@ -71,3 +71,11 @@ export function getNames(task: Task): string[] {
 
   return names
 }
+
+export function getFullName(task: Task, separator = ' > '): string {
+  return getNames(task).join(separator)
+}
+
+export function getTestName(task: Task, separator = ' > '): string {
+  return getNames(task).slice(1).join(separator)
+}
