+++ packages/vitest/src/api/setup.ts
@@ -6,11 +6,13 @@ import type { WebSocket } from 'ws'
 import { WebSocketServer } from 'ws'
 import type { ViteDevServer } from 'vite'
 import type { File, TaskResultPack } from '@vitest/runner'
+import { isPrimitive, noop } from '@vitest/utils'
 import { API_PATH } from '../constants'
 import type { Vitest } from '../node/core'
 import type { Awaitable, ModuleGraphData, UserConsoleLog } from '../types/general'
 import type { Reporter } from '../node/types/reporter'
-import { getModuleGraph, isPrimitive, noop, stringifyReplace } from '../utils'
+import { getModuleGraph } from '../utils/graph'
+import { stringifyReplace } from '../utils/serialization'
 import { parseErrorStacktrace } from '../utils/source-map'
 import type { SerializedTestSpecification } from '../runtime/types/utils'
 import type {
