+++ packages/vitest/src/node/cache/index.ts
@@ -1,5 +1,5 @@
 import { resolve } from 'pathe'
-import { slash } from '../../utils'
+import { slash } from '@vitest/utils'
 import { hash } from '../hash'
 import { FilesStatsCache } from './files'
 import { ResultsCache } from './results'
