+++ packages/vitest/src/node/logger.ts
@@ -5,8 +5,8 @@ import c from 'tinyrainbow'
 import { parseErrorStacktrace } from '@vitest/utils/source-map'
 import type { Task } from '@vitest/runner'
 import type { ErrorWithDiff } from '@vitest/utils'
+import { toArray } from '@vitest/utils'
 import type { TypeCheckError } from '../typecheck/typechecker'
-import { toArray } from '../utils'
 import { highlightCode } from '../utils/colors'
 import { divider } from './reporters/renderers/utils'
 import { RandomSequencer } from './sequencers/RandomSequencer'
