+++ packages/vitest/src/node/plugins/cssEnabler.ts
@@ -1,8 +1,8 @@
 import { relative } from 'pathe'
 import type { Plugin as VitePlugin } from 'vite'
+import { toArray } from '@vitest/utils'
 import { generateCssFilenameHash } from '../../integrations/css/css-modules'
 import type { CSSModuleScopeStrategy, ResolvedConfig } from '../types/config'
-import { toArray } from '../../utils'
 
 const cssLangs = '\\.(?:css|less|sass|scss|styl|stylus|pcss|postcss)(?:$|\\?)'
 const cssLangRE = new RegExp(cssLangs)
