+++ packages/vitest/src/node/plugins/index.ts
@@ -1,13 +1,12 @@
 import type { UserConfig as ViteConfig, Plugin as VitePlugin } from 'vite'
 import { relative } from 'pathe'
-import { configDefaults, coverageConfigDefaults } from '../../defaults'
-import type { ResolvedConfig, UserConfig } from '../types/config'
 import {
   deepMerge,
   notNullish,
-  removeUndefinedValues,
   toArray,
-} from '../../utils'
+} from '@vitest/utils'
+import { configDefaults, coverageConfigDefaults } from '../../defaults'
+import type { ResolvedConfig, UserConfig } from '../types/config'
 import { resolveApiServerConfig } from '../config/resolveConfig'
 import { Vitest } from '../core'
 import { generateScopedClassName } from '../../integrations/css/css-modules'
@@ -265,4 +264,14 @@ export async function VitestPlugin(
     VitestOptimizer(),
     NormalizeURLPlugin(),
   ].filter(notNullish)
+}
+function removeUndefinedValues<T extends Record<string, any>>(
+  obj: T,
+): T {
+  for (const key in Object.keys(obj)) {
+    if (obj[key] === undefined) {
+      delete obj[key]
+    }
+  }
+  return obj
 }
