+++ packages/vitest/src/node/pools/forks.ts
@@ -4,11 +4,12 @@ import EventEmitter from 'node:events'
 import { Tinypool } from 'tinypool'
 import type { TinypoolChannel, Options as TinypoolOptions } from 'tinypool'
 import { createBirpc } from 'birpc'
+import { resolve } from 'pathe'
 import type { PoolProcessOptions, ProcessPool, RunWithFiles } from '../pool'
 import type { WorkspaceProject } from '../workspace'
 import { envsOrder, groupFilesByEnv } from '../../utils/test-helpers'
 import { wrapSerializableConfig } from '../../utils/config-helpers'
-import { groupBy, resolve } from '../../utils'
+import { groupBy } from '../../utils/base'
 import type { SerializedConfig } from '../types/config'
 import type { RunnerRPC, RuntimeRPC } from '../../types/rpc'
 import type { Vitest } from '../core'
