+++ packages/vitest/src/node/pools/threads.ts
@@ -6,7 +6,7 @@ import Tinypool from 'tinypool'
 import { resolve } from 'pathe'
 import type { PoolProcessOptions, ProcessPool, RunWithFiles } from '../pool'
 import { envsOrder, groupFilesByEnv } from '../../utils/test-helpers'
-import { AggregateError, groupBy } from '../../utils/base'
+import { groupBy } from '../../utils/base'
 import type { WorkspaceProject } from '../workspace'
 import type { SerializedConfig } from '../types/config'
 import type { RunnerRPC, RuntimeRPC } from '../../types/rpc'
