+++ packages/vitest/src/node/pools/typecheck.ts
@@ -1,9 +1,9 @@
 import type { DeferPromise } from '@vitest/utils'
 import { createDefer } from '@vitest/utils'
+import { hasFailed } from '@vitest/runner/utils'
 import type { TypecheckResults } from '../../typecheck/typechecker'
 import { Typechecker } from '../../typecheck/typechecker'
 import { groupBy } from '../../utils/base'
-import { hasFailed } from '../../utils/tasks'
 import type { Vitest } from '../core'
 import type { ProcessPool, WorkspaceSpec } from '../pool'
 import type { WorkspaceProject } from '../workspace'
