+++ packages/vitest/src/node/pools/vmForks.ts
@@ -8,7 +8,6 @@ import Tinypool from 'tinypool'
 import { rootDir } from '../../paths'
 import type { PoolProcessOptions, ProcessPool, RunWithFiles } from '../pool'
 import { groupFilesByEnv } from '../../utils/test-helpers'
-import { AggregateError } from '../../utils/base'
 import type { WorkspaceProject } from '../workspace'
 import { getWorkerMemoryLimit, stringToBytes } from '../../utils/memory-limit'
 import { wrapSerializableConfig } from '../../utils/config-helpers'
