+++ packages/vitest/src/node/pools/vmThreads.ts
@@ -7,7 +7,6 @@ import Tinypool from 'tinypool'
 import { rootDir } from '../../paths'
 import type { PoolProcessOptions, ProcessPool, RunWithFiles } from '../pool'
 import { groupFilesByEnv } from '../../utils/test-helpers'
-import { AggregateError } from '../../utils/base'
 import type { WorkspaceProject } from '../workspace'
 import { getWorkerMemoryLimit, stringToBytes } from '../../utils/memory-limit'
 import type { ResolvedConfig, SerializedConfig } from '../types/config'
