+++ packages/vitest/src/node/reporters/benchmark/table/index.ts
@@ -2,11 +2,12 @@ import fs from 'node:fs'
 import c from 'tinyrainbow'
 import * as pathe from 'pathe'
 import type { File, TaskResultPack } from '@vitest/runner'
+import { getTasks } from '@vitest/runner/utils'
 import type { UserConsoleLog } from '../../../../types/general'
 import { BaseReporter } from '../../base'
-import { getFullName, getTasks } from '../../../../utils'
 import { getStateSymbol } from '../../renderers/utils'
 import type { BenchmarkResult } from '../../../../runtime/types/benchmark'
+import { getFullName } from '../../../../utils/tasks'
 import {
   type TableRendererOptions,
   createTableRenderer,
