+++ packages/vitest/src/node/reporters/benchmark/table/tableRender.ts
@@ -2,11 +2,12 @@ import { stripVTControlCharacters } from 'node:util'
 import c from 'tinyrainbow'
 import cliTruncate from 'cli-truncate'
 import type { Task } from '@vitest/runner'
-import { getTests, notNullish } from '../../../../utils'
-import { F_RIGHT } from '../../../../utils/figures'
+import { notNullish } from '@vitest/utils'
+import { getTests } from '@vitest/runner/utils'
 import type { Logger } from '../../../logger'
 import { getCols, getStateSymbol } from '../../renderers/utils'
 import type { BenchmarkResult } from '../../../../runtime/types/benchmark'
+import { F_RIGHT } from '../../renderers/figures'
 import type { FlatBenchmarkReport } from '.'
 
 export interface TableRendererOptions {
