+++ packages/vitest/src/node/reporters/json.ts
@@ -2,7 +2,7 @@ import { existsSync, promises as fs } from 'node:fs'
 import { dirname, resolve } from 'pathe'
 import type { File, Suite, TaskMeta, TaskState } from '@vitest/runner'
 import type { SnapshotSummary } from '@vitest/snapshot'
-import { getSuites, getTests } from '../../utils'
+import { getSuites, getTests } from '@vitest/runner/utils'
 import { getOutputFile } from '../../utils/config-helpers'
 import type { Reporter } from '../types/reporter'
 import type { Vitest } from '../core'
