+++ packages/vitest/src/node/reporters/renderers/figures.ts
