+++ packages/vitest/src/node/reporters/renderers/listRenderer.ts
@@ -2,11 +2,12 @@ import { stripVTControlCharacters } from 'node:util'
 import c from 'tinyrainbow'
 import cliTruncate from 'cli-truncate'
 import type { SuiteHooks, Task } from '@vitest/runner'
-import { getTests, notNullish } from '../../../utils'
-import { F_RIGHT } from '../../../utils/figures'
+import { getTests } from '@vitest/runner/utils'
+import { notNullish } from '@vitest/utils'
 import type { Logger } from '../../logger'
 import type { VitestRunMode } from '../../types/config'
 import type { Benchmark, BenchmarkResult } from '../../../runtime/types/benchmark'
+import { F_RIGHT } from './figures'
 import {
   formatProjectName,
   getCols,
