+++ packages/vitest/src/node/reporters/renderers/utils.ts
@@ -3,7 +3,7 @@ import { basename, dirname, isAbsolute, relative } from 'pathe'
 import c from 'tinyrainbow'
 import type { SuiteHooks, Task } from '@vitest/runner'
 import type { SnapshotSummary } from '@vitest/snapshot'
-import { slash } from '../../../utils/base'
+import { slash } from '@vitest/utils'
 import {
   F_CHECK,
   F_CROSS,
@@ -12,7 +12,7 @@ import {
   F_DOWN_RIGHT,
   F_LONG_DASH,
   F_POINTER,
-} from '../../../utils/figures'
+} from './figures'
 
 export const spinnerMap = new WeakMap<Task, () => string>()
 export const hookSpinnerMap = new WeakMap<Task, Map<string, () => string>>()
