+++ packages/vitest/src/node/watch-filter.ts
@@ -3,7 +3,7 @@ import type { Writable } from 'node:stream'
 import { stripVTControlCharacters } from 'node:util'
 import c from 'tinyrainbow'
 import { createDefer } from '@vitest/utils'
-import { stdout as getStdout } from '../utils'
+import { stdout as getStdout } from '../utils/base'
 
 const MAX_RESULT_COUNT = 10
 const SELECTION_MAX_INDEX = 7
