+++ packages/vitest/src/runtime/runners/test.ts
@@ -12,13 +12,14 @@ import type {
 } from '@vitest/runner'
 import type { ExpectStatic } from '@vitest/expect'
 import { GLOBAL_EXPECT, getState, setState } from '@vitest/expect'
+import { getNames, getTestName, getTests } from '@vitest/runner/utils'
 import { getSnapshotClient } from '../../integrations/snapshot/chai'
 import { vi } from '../../integrations/vi'
-import { getNames, getTestName, getTests, getWorkerState } from '../../utils'
 import { createExpect } from '../../integrations/chai/index'
 import type { SerializedConfig } from '../config'
 import type { VitestExecutor } from '../execute'
 import { rpc } from '../rpc'
+import { getWorkerState } from '../utils'
 
 export class VitestTestRunner implements VitestRunner {
   private snapshotClient = getSnapshotClient()
