+++ packages/vitest/src/runtime/setup-node.ts
@@ -3,11 +3,12 @@ import util from 'node:util'
 import timers from 'node:timers'
 import { installSourcemapsSupport } from 'vite-node/source-map'
 import { KNOWN_ASSET_TYPES } from 'vite-node/constants'
-import { getSafeTimers, getWorkerState } from '../utils'
+import { getSafeTimers } from '@vitest/utils'
 import * as VitestIndex from '../public/index'
 import { expect } from '../integrations/chai'
 import { resolveSnapshotEnvironment } from '../integrations/snapshot/environments/resolveSnapshotEnvironment'
 import type { ResolvedTestEnvironment } from '../types/environment'
+import { getWorkerState } from './utils'
 import { setupCommonEnv } from './setup-common'
 import type { VitestExecutor } from './execute'
 import type { SerializedConfig } from './config'
