+++ packages/vitest/src/runtime/utils.ts
@@ -1,8 +1,13 @@
+import type { ModuleCacheMap } from 'vite-node/client'
+
+import { getSafeTimers } from '@vitest/utils'
 import type { WorkerGlobalState } from '../types/worker'
+
+const NAME_WORKER_STATE = '__vitest_worker__'
 
 export function getWorkerState(): WorkerGlobalState {
   // @ts-expect-error untyped global
-  const workerState = globalThis.__vitest_worker__
+  const workerState = globalThis[NAME_WORKER_STATE]
   if (!workerState) {
     const errorMsg
       = 'Vitest failed to access its internal state.'
@@ -16,7 +21,7 @@ export function getWorkerState(): WorkerGlobalState {
 }
 
 export function provideWorkerState(context: any, state: WorkerGlobalState) {
-  Object.defineProperty(context, '__vitest_worker__', {
+  Object.defineProperty(context, NAME_WORKER_STATE, {
     value: state,
     configurable: true,
     writable: true,
@@ -29,4 +34,60 @@ export function provideWorkerState(context: any, state: WorkerGlobalState) {
 export function getCurrentEnvironment(): string {
   const state = getWorkerState()
   return state?.environment.name
+}
+
+export function isChildProcess(): boolean {
+  return typeof process !== 'undefined' && !!process.send
+}
+
+export function setProcessTitle(title: string) {
+  try {
+    process.title = `node (${title})`
+  }
+  catch {}
+}
+
+export function resetModules(modules: ModuleCacheMap, resetMocks = false) {
+  const skipPaths = [
+    // Vitest
+    /\/vitest\/dist\//,
+    /\/vite-node\/dist\//,
+    // yarn's .store folder
+    /vitest-virtual-\w+\/dist/,
+    // cnpm
+    /@vitest\/dist/,
+    // don't clear mocks
+    ...(!resetMocks ? [/^mock:/] : []),
+  ]
+  modules.forEach((mod, path) => {
+    if (skipPaths.some(re => re.test(path))) {
+      return
+    }
+    modules.invalidateModule(mod)
+  })
+}
+
+function waitNextTick() {
+  const { setTimeout } = getSafeTimers()
+  return new Promise(resolve => setTimeout(resolve, 0))
+}
+
+export async function waitForImportsToResolve() {
+  await waitNextTick()
+  const state = getWorkerState()
+  const promises: Promise<unknown>[] = []
+  let resolvingCount = 0
+  for (const mod of state.moduleCache.values()) {
+    if (mod.promise && !mod.evaluated) {
+      promises.push(mod.promise)
+    }
+    if (mod.resolving) {
+      resolvingCount++
+    }
+  }
+  if (!promises.length && !resolvingCount) {
+    return
+  }
+  await Promise.allSettled(promises)
+  await waitForImportsToResolve()
 }
