+++ packages/vitest/src/runtime/worker.ts
@@ -2,8 +2,8 @@ import { pathToFileURL } from 'node:url'
 import { workerId as poolId } from 'tinypool'
 import { ModuleCacheMap } from 'vite-node/client'
 import { loadEnvironment } from '../integrations/env/loader'
-import { isChildProcess, setProcessTitle } from '../utils/base'
 import type { ContextRPC, WorkerGlobalState } from '../types/worker'
+import { isChildProcess, setProcessTitle } from './utils'
 import { setupInspect } from './inspector'
 import { createRuntimeRpc, rpcDone } from './rpc'
 import type { VitestWorker } from './workers/types'
