+++ packages/vitest/src/typecheck/assertType.ts
@@ -1,7 +1,5 @@
-function noop() {}
-
 export interface AssertType {
   <T>(value: T): void
 }
 
-export const assertType: AssertType = noop
+export const assertType: AssertType = function assertType() {}
