+++ packages/vitest/src/typecheck/typechecker.ts
@@ -7,7 +7,7 @@ import type { RawSourceMap } from '@ampproject/remapping'
 import type { ParsedStack } from '@vitest/utils'
 import type { File, Task, TaskResultPack, TaskState } from '@vitest/runner'
 import { x } from 'tinyexec'
-import { getTasks } from '../utils'
+import { getTasks } from '@vitest/runner/utils'
 import type { WorkspaceProject } from '../node/workspace'
 import type { Awaitable } from '../types/general'
 import type { Vitest } from '../node/core'
